FT Ajout de tests (#77)
* FT Ajout de tests

+++ src/main/kotlin/com/dash/security/response/JwtResponse.kt
@@ -2,7 +2,7 @@ package com.dash.security.response
 
 class JwtResponse(
     var accessToken: String,
-    var id: Long?,
+    var id: Int?,
     var username: String?,
     var email: String?,
     val roles: List<String>
